Готовы задачи с 1 по 3
