--- lesson6/task4/Program.cs
@@ -1,11 +1,0 @@
-﻿using System;
-class Program
-{
-    static void Main(string[] args)
-    {
-        //Выполнил Ищенко Дмитрий
-        //**Считайте файл различными способами.Смотрите “Пример записи файла различными способами”. 
-        //Создайте методы, которые возвращают массив byte(FileStream, BufferedStream), строку для StreamReader и массив int для BinaryReader.
-    }
-}
-
